+++ server/autotest_server/testers/custom/custom_tester.py
@@ -4,9 +4,9 @@ from ..specs import TestSpecs
 
 
 class CustomTester(Tester):
-    def __init__(self, specs: TestSpecs) -> None:
+    def __init__(self, specs: TestSpecs, resource_settings: list[tuple[int, tuple[int, int]]] | None = None) -> None:
         """Initialize a CustomTester"""
-        super().__init__(specs, test_class=None)
+        super().__init__(specs, test_class=None, resource_settings=resource_settings)
 
     @Tester.run_decorator
     def run(self) -> None: