+++ server/autotest_server/testers/jupyter/jupyter_tester.py
@@ -58,13 +58,14 @@ class JupyterTester(Tester):
         self,
         specs: TestSpecs,
         test_class: Type[JupyterTest] = JupyterTest,
+        resource_settings: list[tuple[int, tuple[int, int]]] | None = None,
     ):
         """
         Initialize a jupyter tester using the specifications in specs.
 
         This tester will create tests of type test_class.
         """
-        super().__init__(specs, test_class)
+        super().__init__(specs, test_class, resource_settings=resource_settings)
 
     @staticmethod
     def _run_jupyter_tests(test_file: str) -> List[Dict]: