+++ server/install.py
@@ -62,6 +62,11 @@ def create_workspace():
     os.makedirs(config["workspace"], exist_ok=True)
 
 
+def create_worker_log_dir():
+    _print(f'creating worker log directory at {config["worker_log_dir"]}')
+    os.makedirs(config["worker_log_dir"], exist_ok=True)
+
+
 def install_all_testers():
     settings = install_testers()
     skeleton_file = os.path.join(os.path.dirname(os.path.realpath(__file__)), "autotest_server", "schema_skeleton.json")
@@ -76,6 +81,7 @@ def install():
     check_dependencies()
     check_users_exist()
     create_workspace()
+    create_worker_log_dir()
     install_all_testers()